add single beam as data unit (#15)
This PR adds support for a "single beam" data unit in the OMNIC reader by updating the header documentation and introducing handling for a new key (15). The changes include updating the data unit list in the header comment, adding a new conditional branch for key 15, and improving the log message for unrecognized data.